refactor(agent): extract agent-template helpers, descriptors, and rows (phase 1)
Pure, verbatim move out of the 2070-line AgentTemplateControl into a new
`agentTemplate/` folder — no logic change:

- agentTemplateUtils.ts — enumLabel, countSummary, cloneItem
- itemDescriptors.tsx — ItemDescriptor + describeTool/Mcp/Skill/Instruction and
  their classifiers (toolName, isFunctionTool, isStaticSkill, isEmbedRefSkill,
  mdPreview, …)
- ItemRow.tsx — ItemAvatar, ItemRow, InstructionsFileRow

Orchestrator re-imports them; dropped the now-dead phosphor/toolUtils imports.
2070 → 1628 lines. tsc + lint clean.
